[frontend] StixNestedRefRelationshipEdition in tsx
